Create Dashboard for Recrutiment Process and selection process Dashboard.
