+++ src/webparts/recruitmentApp/Screens/RecuritmentProcess/CommanFilter.ts
@@ -31,6 +31,18 @@ export const getRecruitmentFiltersByTab = (props: any, tab: string) => {
       );
       break;
 
+    case TabName.InterviewQuestion:
+      filters.push(
+        { FilterKey: "StatusId", Operator: "in", FilterValue: [StatusId.PendingwithHRandLMtocreateinterviewQuestion, StatusId.PendingwithLMcreateDisqualificationQuestion] },
+      );
+      break;
+
+    case TabName.ReviewScorecard:
+      filters.push(
+        { FilterKey: "StatusId", Operator: "eq", FilterValue: StatusId.RecruitmentInProgress },
+      );
+      break;
+
     case TabName.ReviewProfile:
       filters.push(
         { FilterKey: "StatusId", Operator: "eq", FilterValue: StatusId.RecruitmentInProgress },
