+++ src/webparts/recruitmentApp/Services/InitiateOfferLetter/IOfferLetterService.ts
@@ -107,7 +107,18 @@ export type GetDOTAfricaCF = {
     Natioality: string;
 }
 
+export type tabcountBGV = {
+    BGVCount: number;
+    LabourHireCount: number;
+    KCSACount: number;
+};
+
 export type IOfferLetterService = {
+    GetBVGCandidateCount(
+        Filter: any[],
+        Conditions: any,
+        CurrentRoleID: number[]
+    ): Promise<ApiResponse<tabcountBGV>>;
     UploadCandidateDocument(
         DocumentName: DocumentName,
         AttachFile: IDocFiles[],
